+++ lib/screens.dart
@@ -15,3 +15,4 @@ export 'src/screens/lit_sign_up_screen.dart';
 export 'src/screens/photosensitive_disclaimer_screen.dart';
 export 'src/screens/confirm_age_screen.dart';
 export 'src/screens/lit_startup_screen.dart';
+export 'src/screens/lit_privacy_disclaimer_screen.dart';