+++ lib/screens.dart
@@ -13,6 +13,6 @@ export 'src/screens/lit_offline_app_disclaimer_screen.dart';
 export 'src/screens/lit_credits_screen.dart';
 export 'src/screens/lit_sign_up_screen.dart';
 export 'src/screens/photosensitive_disclaimer_screen.dart';
-export 'src/screens/confirm_age_screen.dart';
+export 'src/screens/lit_confirm_age_screen.dart';
 export 'src/screens/lit_startup_screen.dart';
 export 'src/screens/lit_privacy_disclaimer_screen.dart';